Textured terrain
Managed to load terrain textures into a single terrain texture, it is sampled by precalculating the uv's and storing them in a dict that gets used to set the uv's in create_face the shader then uses a mipmap_offset based on texture size multiplied by zoom level to find the right mipmap subtexture
Currently the texture loading is quite slow so I'll look to optimize it.
It also picks a texture per quad which should be per triangle.
The textures are spread out across multiple tiles in the base game so when setting the uv-values I should use the x and y coordinates to pick a sub-square of the uv.
it also appears that some rounding causes the edges of textures to be from a different texture
Next step after adressing these issues would be to look into blending the edges of textures and smoothening out the geometry