fixed a loading issue on dbpf files and added water
Due to a mistake where the loading process was attempting to print a string with dbpf.filename, some files wern't loading, this is because the dpbf files didn't have a filename variable and so the thread loading the file would load its subfiles untill it encountered one of the print calls. Fixed by replacing the call to filename with path. I stumbled upon this because it wouldn't load the water textures, but I think this is what sh was strugling with in loading the xml files for the ui. In addition I implemented a way to limit beaches to only near the shores and made the watersurface draw the texture with transparency around the shores.